+++ CHANGELOG.md
@@ -8,9 +8,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 <!-- next-header -->
 ## [Unreleased] - ReleaseDate
 
+### Added
+- Derive `Copy`, `PartialEq`, `Eq` for `SlaveAddr`.
+
 ### Changed
 - Update to `embedded-hal` 1.0.
 - The MSRV is now 1.62.0.
+
+### Removed
+- `Default` derive for device structs.
 
 ## [0.2.1] - 2020-08-13